requirements.txt
# === Основные ===
aiogram==3.16.0
aiohttp==3.11.5
aiofiles==24.1.0

# === База данных ===
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# === Кэширование ===
redis==5.0.7
aioredis==2.0.1

# === Валидация и конфигурация ===
pydantic==2.10.3
pydantic-settings==2.6.0

# === Фоновые задачи ===
apscheduler==3.10.4

# === Excel и данные ===
openpyxl==3.1.5
pandas==2.2.3
xlsxwriter==3.2.0

# === Работа с файлами ===
python-magic==0.4.27
Pillow==11.0.0

# === Утилиты ===
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
pathvalidate==3.2.3
phonenumbers==8.13.45

# === Безопасность ===
bcrypt==4.2.0
cryptography==43.0.0

# === Логирование ===
structlog==24.4.0

# === Разработка ===
black==24.1.1
isort==5.13.2
pytest==8.0.0
pytest-asyncio==0.23.5